+++ src/main/java/org/jabref/gui/preferences/customentrytypes/FieldViewModel.java
@@ -70,6 +70,10 @@ public class FieldViewModel {
         return priorityProperty.getValue();
     }
 
+    public void setField(String field) {
+        this.fieldName.setValue(field);
+    }
+
     public BibField toBibField() {
         return new BibField(field, priorityProperty.getValue());
     }